Discovered a start-up bug that needed to be fixed. In the U8G_COM_MSG_INIT message handler I wasn't calling the routine to set the pinModes - This meant that it worked if something else had set the appropriate pins to output but not from a cold start.
I also added some explicit casts to get rid of the C4244 compiler warning that was annoying me.

+++ csrc/u8g_font.c
@@ -997,7 +997,7 @@ void u8g_UpdateRefHeight(u8g_t *u8g)
     ls *= u8g->font_line_spacing_factor;
     ls >>= 6;
   }
-  u8g->line_spacing = ls;
+  u8g->line_spacing = (uint8_t)ls;
 }
 
 void u8g_SetFontRefHeightText(u8g_t *u8g)
